+gain, offset

+++ term/mainframe.hpp
@@ -50,5 +50,5 @@ private:
     std::vector<wxString> history;
     int histpos;
     int addr;
-    int values[8];
+    float values[8];
 };
